Separated Domain and Recipe controllers
